+++ src/Constants/NamingSchemes.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SebastiaanLuca\AutoMorphMap\Constants;
+
+class NamingSchemes
+{
+    /**
+     * @var string
+     */
+    public const SINGULAR_TABLE_NAME = 'table.singular';
+
+    /**
+     * @var string
+     */
+    public const TABLE_NAME = 'table';
+
+    /**
+     * @var string
+     */
+    public const CLASS_BASENAME = 'class.basename';
+}