+++ tests/resources/cache.php
@@ -1,4 +1,4 @@
 <?php return array (
-  0 => 'App\\SomethingCached',
-  1 => 'App\\Models\\CachedUser',
+  0 => 'App\\Models\\SomethingInherited',
+  1 => 'MyPackage\\Models\\Sub\\Package',
 );